Modified Liquid code to allow image links and added image files for image links.
Because image links were defaulting to zooming in on the image rather than redirecting to the href in the 'a' tag, I deleted the inner part of the if statement for the Liquid code. To reimplement the image-link class (that in turn works with various CSS formatting), I need to go back into the _includes/refactor-content.html file and modify it.